Refactors test_routes as a separate module
Fixes failing algolia test where mocker was mocking the wrong method
