+++ tests/conftest.py
@@ -244,7 +244,7 @@ def fake_algolia_unreachable_host(mocker):
 
     mocker.patch('algoliasearch.search_index.SearchIndex.search',
                  side_effect=algolia_exception)
-    mocker.patch('algoliasearch.search_index.SearchIndex.save_object',
+    mocker.patch('algoliasearch.search_index.SearchIndex.save_objects',
                  side_effect=algolia_exception)
     mocker.patch('algoliasearch.search_index.SearchIndex.partial_update_object',
                  side_effect=algolia_exception)
@@ -260,7 +260,7 @@ def fake_algolia_exception(mocker):
 
     mocker.patch('algoliasearch.search_index.SearchIndex.search',
                  side_effect=algolia_exception)
-    mocker.patch('algoliasearch.search_index.SearchIndex.save_object',
+    mocker.patch('algoliasearch.search_index.SearchIndex.save_objects',
                  side_effect=algolia_exception)
     mocker.patch('algoliasearch.search_index.SearchIndex.partial_update_object',
                  side_effect=algolia_exception)
